# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Shalomut Map API (מפת שלומות)",
    "version": "1.0.0",
    "description": "Clean Data Layer & REST API for Shalomut Map — visualizing school staff wellbeing across 8 organic dimensions. Implements ADR-001 (Raw Data & Privacy Threshold Control)."
  },
  "servers": [
    {
      "url": "https://shalomut-map-demo-ui-redesign.vercel.app",
      "description": "Staging Environment"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
...
  "paths": {
...
    "/api/manager/setup": {
      "put": {
        "tags": ["Manager Setup"],
        "summary": "Create or update manager setup",
        "description": "Persists the school profile, survey round, privacy threshold, dates, and background context. Existing records are updated when their IDs are supplied.",
        "security": [
          { "basicAuth": [] }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManagerSetupInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Manager setup saved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "organization": { "$ref": "#/components/schemas/Organization" },
                    "round": { "$ref": "#/components/schemas/SurveyRound" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid manager setup payload",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Organization or survey round not found in the manager scope"
          },
          "500": {
            "description": "Failed to save manager setup",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "503": {
            "description": "Persistent storage is not configured for this deployment"
          }
        }
      }
    },
...
    "/api/rounds/{roundId}": {
      "patch": {
        "tags": ["Rounds"],
        "summary": "Transition a survey round status",
        "security": [
          { "basicAuth": [] }
        ],
        "parameters": [
          {
            "name": "roundId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "4f88af95-0ec4-4872-85ac-67466c3cd8f2"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["status"],
                "properties": {
                  "status": { "$ref": "#/components/schemas/RoundStatus" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Round status updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "round": { "$ref": "#/components/schemas/SurveyRound" }
                  }
                }
              }
            }
          },
          "400": { "description": "Invalid status" },
          "404": { "description": "Survey round not found" },
          "409": { "description": "Status transition is not allowed" },
          "500": { "description": "Failed to update survey round" },
          "503": { "description": "Persistent storage is not configured for this deployment" }
        }
      }
    },
    "/api/rounds/{roundId}/survey-definition": {
      "parameters": [
        {
          "name": "roundId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "example": "4f88af95-0ec4-4872-85ac-67466c3cd8f2"
          }
        }
      ],
      "get": {
        "tags": ["Manager Setup"],
        "summary": "Fetch the saved survey definition",
        "description": "Returns the saved definition or the canonical 24-question definition when the round has not been customized.",
        "security": [
          { "basicAuth": [] }
        ],
        "responses": {
          "200": {
            "description": "Survey definition retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "definition": { "$ref": "#/components/schemas/SurveyDefinition" }
                  }
                }
              }
            }
          },
          "404": { "description": "Survey round not found" }
        }
      },
      "put": {
        "tags": ["Manager Setup"],
        "summary": "Save the survey definition",
        "description": "Persists survey copy and optional questions. The 24 canonical questions must remain enabled, required, unique, and assigned to their canonical dimensions.",
        "security": [
          { "basicAuth": [] }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SurveyDefinition" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Survey definition saved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "definition": { "$ref": "#/components/schemas/SurveyDefinition" }
                  }
                }
              }
            }
          },
          "400": { "description": "Invalid survey definition" },
          "404": { "description": "Survey round not found" },
          "500": { "description": "Failed to save survey definition" },
          "503": { "description": "Persistent storage is not configured for this deployment" }
        }
      }
    },
...
    "/api/rounds/{roundId}/ai-insights": {
      "parameters": [
        {
          "name": "roundId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "example": "4f88af95-0ec4-4872-85ac-67466c3cd8f2"
          }
        }
      ],
      "get": {
        "tags": ["AI Integration"],
        "summary": "Fetch persisted AI insights",
        "security": [
          { "basicAuth": [] }
        ],
        "responses": {
          "200": {
            "description": "Validated, versioned Stone Map result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoneMapResult"
                }
              }
            }
          },
          "404": {
            "description": "AI insights have not been persisted for this round"
          },
          "500": {
            "description": "Failed to fetch AI insights"
          }
        }
      },
      "post": {
        "tags": ["AI Integration"],
        "summary": "Persist the AI service callback",
        "description": "Validates contract version 1.0, route/payload round identity, privacy semantics, and all eight canonical dimensions before persistence. Bearer authentication is required when AI_CALLBACK_SECRET is configured.",
        "security": [
          { "bearerAuth": [] },
          {}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoneMapResult"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Insights persisted successfully"
          },
          "400": {
            "description": "Payload violates the AI analytics contract"
          },
          "401": {
            "description": "Missing or invalid configured callback secret"
          },
          "404": {
            "description": "Survey round not found"
          },
          "500": {
            "description": "Failed to persist AI insights"
          },
          "503": {
            "description": "Persistent storage is not configured for this deployment"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "Manager credential bound server-side to MANAGER_ORGANIZATION_ID."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Shared service secret. Required only when the corresponding environment variable is configured."
      }
    },
    "schemas": {
      "RoundStatus": {
        "type": "string",
        "enum": ["draft", "active", "closed", "archived"],
        "example": "active"
      },
      "WellbeingStatus": {
        "type": "string",
        "enum": ["green", "yellow", "red"],
        "example": "green"
      },
      "WellbeingDimensionId": {
        "type": "string",
        "enum": [
          "self-expression",
          "professional-competence",
          "social-resource",
          "balance",
          "management-support",
          "certainty",
          "organizational-climate",
          "meaning"
        ],
        "example": "self-expression"
      },
...
      "Organization": {
        "type": "object",
        "required": ["id", "name", "city", "schoolType", "totalStaffCount", "createdAt"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string", "example": "בית ספר דרור" },
          "city": { "type": "string", "example": "תל אביב" },
          "schoolType": { "type": "string", "example": "ממלכתי" },
          "totalStaffCount": { "type": "integer", "minimum": 1, "example": 34 },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "RoundBackgroundContext": {
        "type": "object",
        "required": [
          "notes",
          "audience",
          "sicknessDaysThisQuarter",
          "newStaffMembers",
          "studentCount",
          "socioEconomicIndex",
          "classesPerGrade"
        ],
        "properties": {
          "notes": { "type": "string" },
          "audience": { "type": "string", "example": "כלל צוות ההוראה" },
          "sicknessDaysThisQuarter": { "type": "integer", "minimum": 0 },
          "newStaffMembers": { "type": "integer", "minimum": 0 },
          "studentCount": { "type": "integer", "minimum": 0 },
          "socioEconomicIndex": { "type": "integer", "minimum": 1, "maximum": 10 },
          "classesPerGrade": {
            "type": "object",
            "additionalProperties": { "type": "integer", "minimum": 0 }
          }
        }
      },
      "ManagerSetupInput": {
        "type": "object",
        "required": ["organization", "round"],
        "properties": {
          "organization": {
            "type": "object",
            "required": ["name", "city", "schoolType", "totalStaffCount"],
            "properties": {
              "id": { "type": "string", "format": "uuid" },
              "name": { "type": "string" },
              "city": { "type": "string" },
              "schoolType": { "type": "string" },
              "totalStaffCount": { "type": "integer", "minimum": 1 }
            }
          },
          "round": {
            "type": "object",
            "required": ["title", "privacyThreshold", "startDate", "backgroundContext"],
            "properties": {
              "id": { "type": "string", "format": "uuid" },
              "title": { "type": "string" },
              "privacyThreshold": { "type": "integer", "minimum": 10 },
              "startDate": { "type": "string", "format": "date" },
              "endDate": { "type": "string", "format": "date" },
              "backgroundContext": { "$ref": "#/components/schemas/RoundBackgroundContext" }
            }
          }
        }
      },
      "SurveyDefinitionQuestion": {
        "type": "object",
        "required": ["id", "text", "dimensionId", "required", "enabled", "answerMode"],
        "properties": {
          "id": { "type": "string", "example": "q1" },
          "text": { "type": "string" },
          "dimensionId": { "$ref": "#/components/schemas/WellbeingDimensionId" },
          "required": { "type": "boolean" },
          "enabled": { "type": "boolean" },
          "answerMode": { "type": "string", "example": "סקאלת צבעים" }
        }
      },
      "SurveyDefinition": {
        "type": "object",
        "required": [
          "title",
          "audience",
          "estimatedMinutes",
          "minimumResponses",
          "introText",
          "anonymityText",
          "questions"
        ],
        "properties": {
          "title": { "type": "string" },
          "audience": { "type": "string" },
          "estimatedMinutes": { "type": "number", "minimum": 1 },
          "minimumResponses": { "type": "integer", "minimum": 10 },
          "introText": { "type": "string" },
          "anonymityText": { "type": "string" },
          "questions": {
            "type": "array",
            "minItems": 24,
            "items": { "$ref": "#/components/schemas/SurveyDefinitionQuestion" }
          }
        }
      },
      "SurveyRound": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid", "example": "4f88af95-0ec4-4872-85ac-67466c3cd8f2" },
          "organizationId": { "type": "string", "format": "uuid", "example": "c93f8832-8d0d-43cb-b8d2-cdd6e45683af" },
          "title": { "type": "string", "example": "סקר שלומות - מחצית א'" },
          "status": { "$ref": "#/components/schemas/RoundStatus" },
          "shareCode": { "type": "string", "example": "SHALOM-7K2M" },
          "privacyThreshold": { "type": "integer", "example": 10 },
          "startDate": { "type": "string", "format": "date-time" },
          "endDate": { "type": "string", "format": "date-time" },
          "backgroundContext": { "$ref": "#/components/schemas/RoundBackgroundContext" },
          "surveyDefinition": { "$ref": "#/components/schemas/SurveyDefinition" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "StoneMetric": {
        "type": "object",
        "required": ["label", "value"],
        "properties": {
          "label": { "type": "string" },
          "value": { "type": "string" },
          "trend": { "type": "string" }
        }
      },
      "StoneIntervention": {
        "type": "object",
        "required": ["id", "dimensionId", "source", "title", "summary", "actionable_steps"],
        "properties": {
          "id": { "type": "string" },
          "dimensionId": { "$ref": "#/components/schemas/WellbeingDimensionId" },
          "source": { "type": "string" },
          "title": { "type": "string" },
          "summary": { "type": "string" },
          "actionable_steps": {
            "type": "array",
            "items": { "type": "string" }
          }
        }
      },
      "StoneDetail": {
        "type": "object",
        "required": ["dimensionId", "dimensionNameHebrew", "status", "score", "psychologicalInterpretation", "recommendedInterventions", "metrics"],
        "properties": {
          "dimensionId": { "$ref": "#/components/schemas/WellbeingDimensionId" },
          "dimensionNameHebrew": { "type": "string" },
          "status": { "$ref": "#/components/schemas/WellbeingStatus" },
          "score": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          },
          "psychologicalInterpretation": { "type": "string" },
          "recommendedInterventions": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/StoneIntervention" }
          },
          "metrics": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/StoneMetric" }
          }
        }
      },
      "StoneMapResult": {
        "type": "object",
        "required": ["contractVersion", "roundId", "isLocked", "status"],
        "properties": {
          "contractVersion": {
            "type": "string",
            "enum": ["1.0"],
            "example": "1.0"
          },
          "roundId": {
            "type": "string",
            "example": "4f88af95-0ec4-4872-85ac-67466c3cd8f2"
          },
          "processedAt": {
            "type": "string",
            "format": "date-time"
          },
          "isLocked": {
            "type": "boolean"
          },
          "status": {
            "type": "string",
            "enum": ["success", "locked_error", "validation_failed"]
          },
          "errorMessage": {
            "type": "string"
          },
          "overallPsychologicalSummary": {
            "type": "string"
          },
          "stones": {
            "type": "object",
            "minProperties": 8,
            "maxProperties": 8,
            "additionalProperties": {
              "$ref": "#/components/schemas/StoneDetail"
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "Failed to process request" }
        }
      }
    }
  }
}